requirements.txt
Django==1.5.8
MySQL-python==1.2.5
PIL==1.1.7
amqp==1.4.5
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.17
celery==3.1.11
chardet==2.2.1
django-celery==3.1.10
django-ckeditor==4.0.2
django-crispy-forms==1.4.0
django-debug-toolbar==1.2
django-grappelli==2.4.8
django-mongodb-engine==0.5.2
django-redis==3.6.2
django-smtp-ssl==1.0
django-xadmin==0.5.0
djangorestframework==2.3.13
djangotoolbox==1.6.2
ipython==2.0.0
kombu==3.0.16
lxml==3.0.2
mongoengine==0.8.7
pillow==2.5.0
pycrypto==2.6
pymongo==2.7.2
pyreadline==2.0
pytz==2014.2
qrcode==4.0.4
redis==2.9.1
six==1.6.1
sqlparse==0.1.11
suds==0.4
xlrd==0.9.3
